Revert "refactor(agent-status): websocket connector, internal hook tests (#196)"
This reverts commit c1f3bff07d5686003330b04f351f8f1c36a1ab6c.
