+++ README.md
@@ -1,5 +1,3 @@
-# Task complete marker
-
 # 馬可無序 (makamujo)
 MAKA Mujo: an AI‑VTuber
