+++ console/index.ts
@@ -17,21 +17,6 @@ export type ConsoleServer = {
   stop(closeActiveConnections?: boolean): void;
 };
 
-export type AgentStateProvider = () => unknown;
-
-export function startConsoleServer(certPath?: string, keyPath?: string): ConsoleServer;
-export function startConsoleServer(getAgentState: AgentStateProvider, certPath?: string, keyPath?: string): ConsoleServer;
-export function startConsoleServer(
-  arg1?: string | AgentStateProvider,
-  arg2?: string,
-  arg3?: string,
-): ConsoleServer {
-  if (typeof arg1 === "function") {
-    return startConsoleServerImpl(arg1, arg2 ?? consoleCertPath, arg3 ?? consoleKeyPath);
-  }
-  return startConsoleServerImpl(() => ({}), arg1 ?? consoleCertPath, arg2 ?? consoleKeyPath);
-}
-
 /**
  * Build the redirect response used when an access to the outer console server is denied.
  *
@@ -52,7 +37,7 @@ export function createAccessDeniedRedirectResponse(requestURL: URL): Response {
 }
 
 /**
- * Start the console server implementation.
+ * Start the console server.
  *
  * Internally uses two servers:
  * - A loopback server that serves all console routes (including HTML bundling) on 127.0.0.1.
@@ -60,50 +45,22 @@ export function createAccessDeniedRedirectResponse(requestURL: URL): Response {
  *
  * The returned handle exposes only the outer server's URL and a unified `stop()` method.
  *
- * @param getAgentState - Provider that returns the current agent state payload for WebSocket broadcasts.
- * @param certPath - Path to the TLS certificate file.
- * @param keyPath - Path to the TLS private key file.
+ * @param certPath - Path to the TLS certificate file. Defaults to the `CONSOLE_TLS_CERT` env var.
+ * @param keyPath  - Path to the TLS private key file. Defaults to the `CONSOLE_TLS_KEY` env var.
+ * @returns A handle with the outer server's URL and a unified `stop()` method.
  */
-function startConsoleServerImpl(
-  getAgentState: AgentStateProvider,
-  certPath: string,
-  keyPath: string,
-): ConsoleServer {
-  // If the caller requested loopback-only mode, skip TLS certificate checks
-  // since no outer TLS server will be started. Otherwise, fail fast when
-  // TLS assets are missing to avoid confusing runtime errors later.
-  if (process.env.CONSOLE_LOOPBACK_ONLY !== '1') {
-    if (!existsSync(certPath) || !existsSync(keyPath)) {
-      throw new Error(
-        `TLS certificate files not found at the resolved paths. ` +
-        `certPath=${JSON.stringify(certPath)}, keyPath=${JSON.stringify(keyPath)}. ` +
-        `Provide valid certPath/keyPath arguments or set CONSOLE_TLS_CERT and CONSOLE_TLS_KEY env vars to the correct paths.`
-      );
-    }
+export function startConsoleServer(certPath: string = consoleCertPath, keyPath: string = consoleKeyPath): ConsoleServer {
+  // Fail fast if TLS cert/key files are missing before starting any servers.
+  if (!existsSync(certPath) || !existsSync(keyPath)) {
+    throw new Error(
+      `TLS certificate files not found at the resolved paths. ` +
+      `certPath=${JSON.stringify(certPath)}, keyPath=${JSON.stringify(keyPath)}. ` +
+      `Provide valid certPath/keyPath arguments or set CONSOLE_TLS_CERT and CONSOLE_TLS_KEY env vars to the correct paths.`
+    );
   }
 
   const accessLogger = createDailyRotatingJsonLogger(consoleAccessLogPath);
   const errorLogger = createDailyRotatingJsonLogger(consoleErrorLogPath);
-  const consoleWebSocketPath = `${consoleBasePath}api/ws`;
-  const webSocketClients = new Set<WebSocket>();
-  let lastWebSocketAgentStatePayload = "";
-  const agentStateBroadcastInterval = setInterval(() => {
-    if (webSocketClients.size === 0) {
-      return;
-    }
-    const serializedState = JSON.stringify(getAgentState());
-    if (serializedState === lastWebSocketAgentStatePayload) {
-      return;
-    }
-    lastWebSocketAgentStatePayload = serializedState;
-    for (const client of webSocketClients) {
-      try {
-        client.send(serializedState);
-      } catch {
-        webSocketClients.delete(client);
-      }
-    }
-  }, 1_000);
 
   // Loopback console server: binds to 127.0.0.1 only and serves all console routes
   // (including HTML bundling). Not exposed to the public network.
@@ -121,22 +78,6 @@ function startConsoleServerImpl(
   });
 
   const loopbackConsolePort = loopbackServer.port;
-
-  // If explicitly requested, expose only the loopback server and do not attempt
-  // to bind the outer TLS server. This is useful for CI and local E2E runs
-  // where binding privileged ports or providing TLS certs is inconvenient.
-  if (process.env.CONSOLE_LOOPBACK_ONLY === '1') {
-    return {
-      get url() { return new URL(`http://127.0.0.1:${loopbackConsolePort}`); },
-      stop(closeActiveConnections?: boolean) {
-        clearInterval(agentStateBroadcastInterval);
-        for (const client of webSocketClients) {
-          client.close();
-        }
-        loopbackServer.stop(closeActiveConnections);
-      },
-    };
-  }
 
   // Outer console server: exposed publicly on port 443.
   // Checks the client IP against the shared allowlist before proxying to the loopback server.
@@ -193,38 +134,6 @@ function startConsoleServerImpl(
           proxyHeaders.delete('origin');
           proxyHeaders.delete('referer');
 
-          if (requestURL.pathname === consoleWebSocketPath && req.headers.get('upgrade')?.toLowerCase() === 'websocket') {
-            statusCode = 101;
-            const WebSocketPairConstructor = (globalThis as unknown as { WebSocketPair?: new () => unknown }).WebSocketPair;
-            if (WebSocketPairConstructor === undefined) {
-              throw new Error("WebSocketPair is not available in this runtime");
-            }
-            const webSocketPair = new WebSocketPairConstructor() as { [index: number]: any };
-            const serverSocket = webSocketPair[1];
-            if (!serverSocket) {
-              throw new Error("WebSocketPair did not provide a server socket");
-            }
-            serverSocket.accept();
-            serverSocket.addEventListener('open', () => {
-              webSocketClients.add(serverSocket);
-              try {
-                serverSocket.send(JSON.stringify(getAgentState()));
-              } catch {
-                // intentionally ignored
-              }
-            });
-            serverSocket.addEventListener('message', () => {
-              // no-op for client subscriptions
-            });
-            serverSocket.addEventListener('close', () => {
-              webSocketClients.delete(serverSocket);
-            });
-            serverSocket.addEventListener('error', () => {
-              webSocketClients.delete(serverSocket);
-            });
-            return new Response(null, { status: 101, webSocket: webSocketPair[0] } as any);
-          }
-
           const response = await fetch(proxyURL.toString(), {
             method: req.method,
             headers: proxyHeaders,
@@ -265,33 +174,13 @@ function startConsoleServerImpl(
       },
     });
   } catch (err) {
-    // If binding to the outer TLS port fails (commonly because the port is
-    // privileged or already in use), do not abort startup. Instead, fall
-    // back to exposing the loopback server directly so E2E tests and local
-    // development can still access console routes without requiring root.
-    console.error('[WARN] failed to start outer console server, falling back to loopback only:', err);
-    // Create a lightweight proxy object that exposes the loopback server URL
-    // and a stop method compatible with the returned ConsoleServer contract.
-    const fallbackUrl = new URL(`http://127.0.0.1:${loopbackConsolePort}`);
-    return {
-      get url() { return fallbackUrl; },
-      stop(closeActiveConnections?: boolean) {
-        clearInterval(agentStateBroadcastInterval);
-        for (const client of webSocketClients) {
-          client.close();
-        }
-        loopbackServer.stop(closeActiveConnections);
-      },
-    };
+    loopbackServer.stop(true);
+    throw err;
   }
 
   return {
     get url() { return outerServer.url; },
     stop(closeActiveConnections?: boolean) {
-      clearInterval(agentStateBroadcastInterval);
-      for (const client of webSocketClients) {
-        client.close();
-      }
       loopbackServer.stop(closeActiveConnections);
       outerServer.stop(closeActiveConnections);
     },
