--- console/src/AgentStatus/Component.test.tsx
@@ -1,100 +1,0 @@
-import { describe, expect, it } from "bun:test";
-import { renderToStaticMarkup } from "react-dom/server";
-import { AgentStatus, createAgentStateWebSocketUrl } from "./Component";
-import { GameStatusSection } from "./sections/GameStatusSection";
-import { LiveDeliveryStatusSection } from "./sections/LiveDeliveryStatusSection";
-import { MarkovModelStatusSection } from "./sections/MarkovModelStatusSection";
-
-describe("AgentStatus layout", () => {
-    it("uses a wider default max width for the status container", () => {
-        const html = renderToStaticMarkup(<AgentStatus />);
-        expect(html).toContain("max-w-7xl");
-        expect(html).toContain("h-full");
-        expect(html).toContain("min-h-0");
-        expect(html).toContain("grid-rows-[auto_minmax(0,1fr)]");
-    });
-});
-
-describe("AgentStatus category sections (sanity)", () => {
-    it("renders live delivery rows", () => {
-        const html = renderToStaticMarkup(
-            <LiveDeliveryStatusSection
-                liveDeliveryRows={[
-                    {
-                        label: "配信指標",
-                        valueComponent: (
-                            <div>
-                                <p>状態</p>
-                                <p>配信中</p>
-                            </div>
-                        ),
-                    },
-                    { label: "配信URL", value: "https://example.com/live", href: "https://example.com/live" },
-                ]}
-            />,
-        );
-        expect(html).toContain("配信状況");
-        expect(html).toContain("配信指標");
-        expect(html).toContain("状態");
-        expect(html).toContain("配信中");
-        expect(html).toContain("https://example.com/live");
-    });
-
-    it("renders markov model rows", () => {
-        const html = renderToStaticMarkup(
-            <MarkovModelStatusSection
-                markovModelRows={[
-                    { label: "生成N-gram", value: "4-gram" },
-                    { label: "発話内容", value: "テスト発話" },
-                    {
-                        label: "これまでの発話",
-                        valueComponent: (
-                            <ul>
-                                <li>
-                                    <p>テスト発話</p>
-                                    <p>4g</p>
-                                    <button type="button" disabled aria-label="学習の取り消し">↩</button>
-                                </li>
-                            </ul>
-                        ),
-                    },
-                ]}
-            />,
-        );
-        expect(html).toContain("マルコフ連鎖モデルの状態");
-        expect(html).toContain("4-gram");
-        expect(html).toContain("テスト発話");
-        expect(html).toContain("これまでの発話");
-        expect(html).toContain("aria-label=\"学習の取り消し\"");
-    });
-
-    it("renders game rows", () => {
-        const html = renderToStaticMarkup(
-            <GameStatusSection
-                gameRows={[
-                    { label: "現在のゲーム", value: "org.dashnet.orteil/cookieclicker" },
-                    {
-                        label: "ゲーム情報",
-                        valueComponent: (
-                            <ul>
-                                <li>status: idle</li>
-                            </ul>
-                        ),
-                    },
-                ]}
-            />,
-        );
-        expect(html).toContain("ゲームの状態");
-        expect(html).toContain("現在のゲーム");
-        expect(html).toContain("org.dashnet.orteil/cookieclicker");
-        expect(html).toContain("ゲーム情報");
-        expect(html).toContain("<ul");
-    });
-});
-
-describe("createAgentStateWebSocketUrl", () => {
-    it("builds a wss URL from a wss base URL", () => {
-        expect(createAgentStateWebSocketUrl("wss://localhost/console/")).toBe("wss://localhost/console/api/ws");
-        expect(createAgentStateWebSocketUrl("wss://127.0.0.1/console/")).toBe("wss://127.0.0.1/console/api/ws");
-    });
-});
