+++ console/src/agentStatusSections/LiveDeliveryStatusSection.tsx
@@ -1,0 +1,12 @@
+import { AgentStatusSectionCard } from "./AgentStatusSectionCard";
+import type { AgentStatusRow } from "./AgentStatusSectionCard";
+
+export const LIVE_DELIVERY_SECTION_TITLE = "配信状況";
+
+type LiveDeliveryStatusSectionProps = {
+  liveDeliveryRows: AgentStatusRow[]
+};
+
+export const LiveDeliveryStatusSection = ({ liveDeliveryRows }: LiveDeliveryStatusSectionProps) => {
+  return <AgentStatusSectionCard title={LIVE_DELIVERY_SECTION_TITLE} rows={liveDeliveryRows} />;
+};
