+++ scripts/debug-agent-status.ts
@@ -1,9 +1,8 @@
-import { cloneAgentStateResponseMockFixture } from "../tests/fixtures/agentStateResponseMock";
-import { createAgentStatusRows, createAgentStatusSections } from "../console/src/AgentStatus";
+import { createMockAgentStateResponse, createAgentStatusRows, createAgentStatusSections } from "../console/src/AgentStatus";
 
 (async () => {
   try {
-    const state = cloneAgentStateResponseMockFixture();
+    const state = createMockAgentStateResponse();
     const rows = createAgentStatusRows(state as any);
     console.log("ROWS:\n", JSON.stringify(rows, null, 2));
     const sections = createAgentStatusSections(state as any);
