+++ tests/e2e/console/console.test.ts
@@ -1,20 +1,17 @@
 import { expect, test } from "@playwright/test";
 import { spawn } from "child_process";
-import { existsSync, writeFileSync, createWriteStream, mkdirSync } from "fs";
-import { cloneAgentStateResponseMockFixture } from "../../fixtures/agentStateResponseMock";
+import { existsSync, writeFileSync } from "fs";
 
-let CONSOLE_BASE_URL = `https://127.0.0.1`;
+const CONSOLE_BASE_URL = `https://127.0.0.1`;
 const BROADCASTING_BASE_URL = `http://127.0.0.1:7777`;
 const SERVER_STARTUP_TIMEOUT_MS = 15_000;
 const BROWSER_PAGE_LOAD_TIMEOUT_MS = 20_000;
 const EXPECTED_CONSOLE_TITLE = "馬可無序 - 管理コンソール";
 
 let server: ReturnType<typeof spawn> | null = null;
-let outStream: import('fs').WriteStream | null = null;
-let errStream: import('fs').WriteStream | null = null;
 
-const waitForServerReady = async (): Promise<string | null> => {
-  return new Promise<string | null>((resolve, reject) => {
+const waitForServerReady = async () => {
+  return new Promise<void>((resolve, reject) => {
     if (!server) {
       reject(new Error("Server process not started"));
       return;
@@ -25,25 +22,22 @@ const waitForServerReady = async (): Promise<string | null> => {
       return;
     }
 
-    const proc = server!;
-    const stdout = proc.stdout!;
-
     let settled = false;
     let buffer = "";
 
     const cleanup = () => {
       clearTimeout(timeout);
-      stdout.off("data", onData);
-      proc.off("exit", onExit);
+      server?.stdout?.off("data", onData);
+      server?.off("exit", onExit);
     };
 
-    const resolveOnce = (url?: string | null) => {
+    const resolveOnce = () => {
       if (settled) {
         return;
       }
       settled = true;
       cleanup();
-      resolve(url ?? null);
+      resolve();
     };
 
     const rejectOnce = (error: Error) => {
@@ -57,14 +51,8 @@ const waitForServerReady = async (): Promise<string | null> => {
 
     const onData = (chunk: Buffer | string) => {
       buffer += chunk.toString();
-      const marker = "Console running at ";
-      const idx = buffer.indexOf(marker);
-      if (idx >= 0) {
-        const rest = buffer.slice(idx + marker.length);
-        const line = (rest.split(/\r?\n/)[0] || "").trim();
-        resolveOnce(line || null);
-      } else if (buffer.includes("Console running")) {
-        resolveOnce(null);
+      if (buffer.includes("Console running")) {
+        resolveOnce();
       }
     };
 
@@ -76,8 +64,8 @@ const waitForServerReady = async (): Promise<string | null> => {
       rejectOnce(new Error("Server startup timed out"));
     }, SERVER_STARTUP_TIMEOUT_MS);
 
-    stdout.on("data", onData);
-    proc.on("exit", onExit);
+    server.stdout.on("data", onData);
+    server.on("exit", onExit);
   });
 };
 
@@ -88,50 +76,13 @@ test.beforeAll(async ({ request }) => {
 
   server = spawn(
     process.platform === "win32" ? "bun.exe" : "bun",
-    ["index.ts", "--port", "7777"],
+    ["start"],
     {
-      env: { ...process.env, NODE_ENV: "production", CONSOLE_TLS_CERT: process.env.CONSOLE_TLS_CERT, CONSOLE_TLS_KEY: process.env.CONSOLE_TLS_KEY, CONSOLE_LOOPBACK_ONLY: '1' },
       stdio: ["ignore", "pipe", "pipe"],
     },
   );
 
-  // Capture server stdout/stderr to files for debugging when tests fail.
-  try {
-    mkdirSync("./var/test-logs", { recursive: true });
-  } catch {}
-  const ts = Date.now();
-  const outPath = `./var/test-logs/console-server-${ts}.log`;
-  const errPath = `./var/test-logs/console-server-${ts}.err.log`;
-  outStream = createWriteStream(outPath);
-  errStream = createWriteStream(errPath);
-  server.stdout?.pipe(outStream);
-  server.stderr?.pipe(errStream);
-
-  const consoleUrl = await waitForServerReady();
-  if (consoleUrl) {
-    CONSOLE_BASE_URL = consoleUrl.replace(/\/$/, '');
-  }
-
-  // Verify the console base URL is responding before continuing.
-  const start = Date.now();
-  const deadline = start + SERVER_STARTUP_TIMEOUT_MS;
-  let lastErr: Error | null = null;
-  while (Date.now() < deadline) {
-    try {
-      const health = await request.get(`${CONSOLE_BASE_URL}/console/robots.txt`);
-      if (health.ok()) {
-        lastErr = null;
-        break;
-      }
-      lastErr = new Error(`unexpected status ${health.status()}`);
-    } catch (err) {
-      lastErr = err as Error;
-    }
-    await new Promise((r) => setTimeout(r, 250));
-  }
-  if (lastErr) {
-    throw new Error(`Console server not responding at ${CONSOLE_BASE_URL}: ${String(lastErr)}`);
-  }
+  await waitForServerReady();
 
   // Register the test runner's IP as the allowed IP so that subsequent
   // requests to the IP-restricted console server are permitted.
@@ -148,8 +99,6 @@ test.afterAll(() => {
     server.kill();
   }
   server = null;
-  try { outStream?.end(); } catch {}
-  try { errStream?.end(); } catch {}
 });
 
 test.describe("console", () => {
@@ -161,48 +110,22 @@ test.describe("console", () => {
   });
 
   test("responds to GET /console/api/agent-state", async ({ request }) => {
-    // The agent state endpoint has been replaced by a WebSocket stream
-    // (`/console/api/ws`). This legacy REST test is intentionally left
-    // as a no-op to avoid false failures in environments where the REST
-    // endpoint is no longer exposed.
-    expect(true).toBeTruthy();
+    const res = await request.get(`${CONSOLE_BASE_URL}/console/api/agent-state`);
+    expect(res.ok()).toBeTruthy();
+    const data = await res.json();
+    expect(data).toHaveProperty("niconama");
   });
 
   test("renders the console app in a browser", async ({ page }) => {
-    // Load the console in mock mode so the UI renders deterministic agent
-    // state without relying on the server or WebSocket timing.
     await page.goto(`${CONSOLE_BASE_URL}/console/?agentStateMock=1`, { waitUntil: "domcontentloaded", timeout: BROWSER_PAGE_LOAD_TIMEOUT_MS });
     expect(await page.title()).toContain(EXPECTED_CONSOLE_TITLE);
     const rootElement = await page.$("#root");
     expect(rootElement).not.toBeNull();
     await expect(page.getByRole("heading", { name: "馬可無序" })).toBeVisible();
     await expect(page.getByText("最終更新:", { exact: false })).toBeVisible();
-
-    // The agent status area may render either an empty placeholder or the details.
-    // Wait for either to appear so CI runs are more deterministic and provide
-    // clearer diagnostics when the details are not present. If the empty
-    // placeholder appears first, give the app a short additional grace period
-    // to populate the details (e.g., when a WebSocket message arrives slightly
-    // after initial render) before failing.
-    const detailsLocator = page.getByTestId("agent-status-details");
-    const emptyLocator = page.getByTestId("agent-status-empty");
-    // Wait up to the browser page load timeout for the details to appear.
-    await detailsLocator.waitFor({ timeout: BROWSER_PAGE_LOAD_TIMEOUT_MS });
-
-    if (await emptyLocator.count() > 0) {
-      // Give the app an extra moment for live updates (WebSocket) to populate
-      // the details before failing the test.
-      try {
-        await detailsLocator.waitFor({ timeout: 5_000 });
-      } catch {
-        const html = await page.content();
-        throw new Error(`Agent status empty in test; page snapshot:\n${html}`);
-      }
-    }
-
-    await expect(detailsLocator).not.toContainText("話せる状態");
-    await expect(detailsLocator).toContainText("生成N-gram");
-    await expect(detailsLocator).toContainText("4-gram");
+    await expect(page.getByTestId("agent-status-details")).not.toContainText("話せる状態");
+    await expect(page.getByTestId("agent-status-details")).toContainText("生成N-gram");
+    await expect(page.getByTestId("agent-status-details")).toContainText("4-gram");
     await expect(page.getByRole("heading", { level: 3, name: "配信状況" })).toBeVisible();
     await expect(page.getByRole("heading", { level: 3, name: "マルコフ連鎖モデルの状態" })).toBeVisible();
     await expect(page.getByRole("heading", { level: 3, name: "ゲームの状態" })).toBeVisible();
